move to tray.py
