+++ openadapt/app/__init__.py
@@ -1,3 +1,3 @@
-from openadapt.app.main import run_app
+from openadapt.app.tray import run_app
 
 run_app()
