+++ openadapt/app/main.py
@@ -1,51 +1,20 @@
-import signal
 import threading
 import base64
 import os
 
-from PIL import Image
 from nicegui import app, ui
-from pystray import Icon, Menu, MenuItem
 
 from openadapt import replay, visualize
-from openadapt.app.cards import recording_prompt, select_import, settings
+from openadapt.app.cards import (
+    recording_prompt,
+    select_import,
+    settings,
+)
 from openadapt.app.util import clear_db, on_export, on_import
 from openadapt.app.objects.console import Console
 
+SERVER = "127.0.0.1:8000/upload"
 FPATH = os.path.dirname(__file__)
-
-app_thread = None
-g_tray = None
-
-
-def new_tray():
-    global g_tray
-
-    def on_exit():
-        global app_thread
-        if app_thread is not None:
-            signal.pthread_kill(app_thread.ident, signal.SIGTERM)
-        os._exit(0)
-
-    def show_app():
-        global app_thread
-        app_thread = threading.Thread(target=start)
-        app_thread.start()
-
-    image = Image.open(f"{FPATH}/assets/logo_dark.png")
-    menu = Menu(
-        MenuItem("Show App", show_app),
-        MenuItem("Exit", on_exit),
-    )
-
-    g_tray = Icon("Pystray", icon=image, menu=menu)
-    return g_tray
-
-
-def run_app():
-    tray = new_tray()
-    # TODO: get tray.run_detached() working for non-blocking solution
-    tray.run()
 
 
 def start():
@@ -117,4 +86,4 @@ def start():
 
 
 if __name__ == "__main__":
-    run_app()
+    start()
