+++ pyproject.toml
@@ -68,4 +68,4 @@ build-backend = "poetry.core.masonry.api"
 visualize = "openadapt.visualize:main"
 record = "openadapt.record:start"
 replay = "openadapt.replay:start"
-app = "openadapt.app.main:run_app"
+app = "openadapt.app.tray:run_app"
